attempted option (3)
I played around with a number of ways to enable a different end-story flow between (a) when a user finishes the story and needs to fill in the feedback , versus (b) when the user visits the reflection browser from the chapter menu (ie later on after completing the story).
i tried:
1. passing in a boolean from the story, but that's the wrong approach because once that variable is set, it is always there (ie cannot distinguish future visits using this)
2. passing in a javascript variable from the chapter page. I think this approach can work but it is onerous.
3. creating a new ui in inkController called StoryEndBrowser and a new component with the same name which just skips the unwanted steps and then linking the "secret chapter" to that new knot/specialTag.

I've implemented (3) for now on this branch  but if @ravern you can think of a neat way to implement (2) let me know.