tweaked ravi story flow
1. Currently the flow is fixed : looking for place > message agent > email uni > choose dorm room or rental room.
What if we gave people a choice to change the flow :
Looking > message agent > choose to accept rental or choose to email uni. If they accept  rental, the flow ends after he gets a property.  If they email uni, then they are presented with  the emails and can choose again if they want to accept the rental still or take the dorm.