Added room to Reflections (chapter and story end) and MCQ Quiz.
To do: add room to savedStates, which is problematic.
We have a technical limit on the in-game choices for the facilitated rooms that will require more time to resolve. For now, we have the MCQ game answers and the student reflections, but not the choices they made during the game. I'll explain next Tuesday - basically we only have one "save game slot" per character, so if a student plays in a room, and then leaves the room and replays the chapter, the new save game (without the room code) will overwrite their homework. the only ways around this require a lot of extra save games to be created which will confuse the game engine itself later on (because we don't have a way for the player to "load the correct save game file").